+++ tests/test_component_blocks.py
@@ -2,28 +2,28 @@ from tests.utils import extract_component_tag
 from tests.main.helpers import render_component_tag
 
 
-def test_component_with_default_block(request):
+def test_component_with_default_block(tetra_request):
     """Tests a simple component with default block"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithDefaultBlock %}content{% /@ %}",
     )
     assert extract_component_tag(content).text == "content"
 
 
-def test_component_with_named_block(request):
+def test_component_with_named_block(tetra_request):
     """Tests a simple component with empty default block (unfilled)"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.simple_component_with_named_block %}{% /@ %}",
     )
     assert extract_component_tag(content).text == ""
 
 
-def test_component_with_named_block_and_content(request):
+def test_component_with_named_block_and_content(tetra_request):
     """Tests a simple component with "foo" block"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithNamedBlock %}"
         "{% block foo %}foo{% endblock %}"
         "{% /@ %}",
@@ -31,19 +31,19 @@ def test_component_with_named_block_and_content(request):
     assert extract_component_tag(content).text == "foo"
 
 
-def test_component_with_named_block_and_default_content(request):
+def test_component_with_named_block_and_default_content(tetra_request):
     """Tests a simple component with "foo" block and default content in it"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithNamedBlockWithContent %}" "{% /@ %}",
     )
     assert extract_component_tag(content).text == "foo"
 
 
-def test_component_with_notexisting_block_and_content(request):
+def test_component_with_notexisting_block_and_content(tetra_request):
     """Tests a simple component with notexisting block filled. Must be ignored."""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithNamedBlock %}"
         "{% block notexisting %}foo{% endblock %}"
         "{% /@ %}",
@@ -51,10 +51,10 @@ def test_component_with_notexisting_block_and_content(request):
     assert extract_component_tag(content).text == ""
 
 
-def test_component_with_named_block_empty(request):
+def test_component_with_named_block_empty(tetra_request):
     """Tests a simple component named"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithNamedBlock %}"
         "{% block foo %}{% endblock %}"
         "{% /@ %}",
@@ -77,11 +77,11 @@ def test_component_with_named_block_empty(request):
 #     assert extract_component(content) == "inside"
 
 
-def test_component_with_2_blocks_unfilled(request):
+def test_component_with_2_blocks_unfilled(tetra_request):
     """Tests a simple component with `foo` and `default` blocks unfilled. Default
     block contains some default content"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWith2Blocks %}"
         "{% block foo %}{% endblock %}"
         "{% /@ %}",
@@ -89,10 +89,10 @@ def test_component_with_2_blocks_unfilled(request):
     assert extract_component_tag(content).text == "default"
 
 
-def test_component_with_2_blocks_partly_filled(request):
+def test_component_with_2_blocks_partly_filled(tetra_request):
     """Tests a simple component with 2 blocks partly filled"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWith2Blocks %}"
         "{% block foo %}bar{% endblock %}"
         "{% /@ %}",
@@ -100,10 +100,10 @@ def test_component_with_2_blocks_partly_filled(request):
     assert extract_component_tag(content).text == "defaultbar"
 
 
-def test_component_with_block_and_default_content_overridden(request):
+def test_component_with_block_and_default_content_overridden(tetra_request):
     """Tests a simple component overridden default content"""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithNamedBlock %}"
         "{% block foo %}overridden{% endblock %}"
         "{% /@ %}",
@@ -111,22 +111,22 @@ def test_component_with_block_and_default_content_overridden(request):
     assert extract_component_tag(content).text == "overridden"
 
 
-def test_component_with_conditional_block_empty(request):
+def test_component_with_conditional_block_empty(tetra_request):
     """Tests a simple component with conditional block that is not rendered,
     as it is empty
     """
     content = render_component_tag(
-        request, "{% @ main.default.SimpleComponentWithConditionalBlock / %}"
+        tetra_request, "{% @ main.default.SimpleComponentWithConditionalBlock / %}"
     )
     assert extract_component_tag(content).text == "always"
 
 
-def test_component_with_conditional_block_filled_empty(request):
+def test_component_with_conditional_block_filled_empty(tetra_request):
     """Tests a simple component with default content, that is overridden with empty
     block.
     """
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithConditionalBlock  %}"
         "{% block foo %}{% endblock %}"
         "{% /@ %}",
@@ -134,11 +134,11 @@ def test_component_with_conditional_block_filled_empty(request):
     assert extract_component_tag(content).decode_contents() == "BEFOREAFTERalways"
 
 
-def test_component_with_conditional_block_filled(request):
+def test_component_with_conditional_block_filled(tetra_request):
     """Tests a simple component with conditional block, filled, with mixed text from
     component and block overrides."""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithConditionalBlock  %}"
         "{% block foo %}foo{% endblock %}"
         "{% /@ %}",
@@ -146,11 +146,11 @@ def test_component_with_conditional_block_filled(request):
     assert extract_component_tag(content).decode_contents() == "BEFOREfooAFTERalways"
 
 
-def test_component_with_conditional_addcontent_block_filled(request):
+def test_component_with_conditional_addcontent_block_filled(tetra_request):
     """Tests a simple component with conditional block, filled, with mixed text from
     component and block overrides."""
     content = render_component_tag(
-        request,
+        tetra_request,
         "{% @ main.default.SimpleComponentWithConditionalBlockAndAdditionalContent %}"
         "{% block foo %}foo{% endblock %}"
         "{% /@ %}",
@@ -158,10 +158,12 @@ def test_component_with_conditional_addcontent_block_filled(request):
     assert extract_component_tag(content).decode_contents() == "BEFOREfooAFTER"
 
 
-def test_component_with_conditional_addcontent_block_filled_and_html_tags(request):
+def test_component_with_conditional_addcontent_block_filled_and_html_tags(
+    tetra_request,
+):
     """Tests a simple component with conditional block, filled, with html tags"""
     content = render_component_tag(
-        request,
+        tetra_request,
         """
 {% @ main.default.SimpleComponentWithConditionalBlockAndAdditionalHtmlContent %}
 {% block foo %}bar{% endblock %}