+++ tests/test_component_context.py
@@ -2,10 +2,10 @@ from tests.utils import extract_component_tag
 from tests.main.helpers import render_component_tag
 
 
-def test_use_extra_context_not_scoped(request):
+def test_use_extra_context_not_scoped(tetra_request):
     """Component may not display outer context vars, if not explicitly included."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithDefaultBlock %}"
         "{{foo}}"
         "{% /@ %}",
@@ -14,11 +14,11 @@ def test_use_extra_context_not_scoped(request):
     assert extract_component_tag(content).text == ""
 
 
-def test_use_extra_context(request):
+def test_use_extra_context(tetra_request):
     """Component must display outer context vars, if explicitly included in
     _extra_context."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithFooContext %}"
         "{{foo}}"
         "{% /@ %}",
@@ -27,11 +27,11 @@ def test_use_extra_context(request):
     assert extract_component_tag(content).text == "bar"
 
 
-def test_use_extra_context_empty(request):
+def test_use_extra_context_empty(tetra_request):
     """Component must not display outer context vars, if explicitly included in
     _extra_context, but var==empty."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithFooContext %}"
         "{{foo}}"
         "{% /@ %}",  # FIXME:KeyError(key)
@@ -40,11 +40,11 @@ def test_use_extra_context_empty(request):
     assert extract_component_tag(content).text == ""
 
 
-def test_use_extra_context_all_empty(request):
+def test_use_extra_context_all_empty(tetra_request):
     """Component must not display outer context vars, if _extra_context == __all__,
     but var==empty."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithExtraContextAll %}"
         "{{foo}}"
         "{% /@ %}",
@@ -53,10 +53,10 @@ def test_use_extra_context_all_empty(request):
     assert extract_component_tag(content).text == ""
 
 
-def test_use_extra_context_all(request):
+def test_use_extra_context_all(tetra_request):
     """Component must display outer context vars, if __all__ in _extra_context."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithExtraContextAll %}"
         "{{foo}}"
         "{% /@ %}",
@@ -68,10 +68,10 @@ def test_use_extra_context_all(request):
 # -------- using template attrs --------
 
 
-def test_use_context_attr(request):
+def test_use_context_attr(tetra_request):
     """context must be available when ctx var explicitly given on template calling"""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithDefaultBlock "
         "context: foo='bar' %}"
         "{{foo}}"
@@ -80,10 +80,10 @@ def test_use_context_attr(request):
     assert extract_component_tag(content).text == "bar"
 
 
-def test_use_context_attr_all(request):
+def test_use_context_attr_all(tetra_request):
     """context must be available when ctx == all on template calling"""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithDefaultBlock "
         "context: __all__ %}"
         "{{foo}}"
@@ -93,10 +93,10 @@ def test_use_context_attr_all(request):
     assert extract_component_tag(content).text == "bar"
 
 
-def test_extra_context(request):
+def test_extra_context(tetra_request):
     """_extra_context must be available automatically."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithExtraContextAll %}"
         "{{foo}}"
         "{% /@ %}",
@@ -105,10 +105,10 @@ def test_extra_context(request):
     assert extract_component_tag(content).text == "bar"
 
 
-def test_context_attr_overrides_extra_context(request):
+def test_context_attr_overrides_extra_context(tetra_request):
     """context given at the template tag must override the outer context."""
     content = render_component_tag(
-        request,
+        tetra_request,
         component_string="{% @ main.default.SimpleComponentWithExtraContextAll "
         "context: foo='nobaz' %}"
         "{{foo}}"